requirements.txt
# Core web
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
PyYAML==6.0.2

# Legacy and plotting deps (UI/templates still use these)
blinker==1.9.0
click==8.1.8
flask==2.3.3
importlib-metadata==8.6.1
itsdangerous==2.2.0
jinja2==3.1.5
markupsafe==3.0.2
numpy==1.21.2
plotly==5.3.1
pyserial==3.5
six==1.17.0
tenacity==9.0.0
werkzeug==2.3.7
zipp==3.21.0

